Crate midly

Overview

midly is a Standard Midi File (SMF) parser focused on speed and flexibility, parsing multi-MB files in tenths of a second.

Usage is as simple as:

use midly::Smf;

let smf = Smf::parse(include_bytes!("../test-asset/Clementi.mid")).unwrap();

for (i, track) in smf.tracks.iter().enumerate() {
    println!("track {} has {} events", i, track.len());
}

The Smf struct is the main type in the crate. See its documentation for the structure of parsed MIDI files.

About lifetimes

The Smf struct is used to store a parsed Standard Midi File (.mid and .midi files). Notice that it has a lifetime parameter, since it stores references to the raw file bytes in order to avoid allocations. For this reason, the byte buffer must be created separately from the Smf structure:

use midly::Smf;

// Load bytes into a buffer
let bytes = include_bytes!("../test-asset/Clementi.mid");

// Parse file in a separate step
let smf = Smf::parse(bytes).unwrap();

When loading a file something similar has to be done:

use std::fs;
use midly::Smf;

// Load bytes into a buffer
let bytes = fs::read("test-asset/Clementi.mid").unwrap();

// Parse bytes in a separate step
let smf = Smf::parse(&bytes).unwrap();

Writing Standard Midi Files

Saving SMF files is as simple as using the Smf::save method:

// Parse file
let bytes = fs::read("test-asset/Clementi.mid").unwrap();
let smf = Smf::parse(&bytes).unwrap();

// Rewrite file
smf.save("test-asset/ClementiRewritten.mid").unwrap();

SMF files can also be written to an arbitrary writer:

let mut in_memory = Vec::new();
smf.write(&mut in_memory).unwrap();

println!("midi file fits in {} bytes!", in_memory.len());

About features

The mode in which the crate works is configurable through the use of cargo features. Two optional features are available: std and strict. Only std is enabled by default.

The std feature

This feature enables the MIDI writer (which uses std::io::Write) and automatic parallelization for the Smf::parse and Smf::parse_with_bytemap functions (through the rayon dependency).

This feature is enabled by default. Disabling this feature with default-features = false will make the crate no_std + alloc.

The strict feature

By default midly will allow corrupted files, by throwing away corrupted events or even entire corrupted tracks. By enabling the strict feature the parser will reject SMF uncompliant files and do additional checking, throwing errors of the kind ErrorKind::Malformed when such a situation arises.

About generics

The Smf type is generic over T, a type implementing TrackRepr. This T indicates how should each track be represented in memory.

The default is Vec<Event>, produced by the Smf::parse method, but there are also two other methods: Smf::parse_with_bytemap and Smf::parse_lazy. Check the documentation for these methods for more information about each.

Parsing raw MIDI streams

The MIDI standard is independent from the Standard Midi File standard, even though the latter depends on the former. This means that raw, non-SMF, MIDI streams exist, for example those generating by MIDI keyboards.

midly provides partial support for parsing these MIDI messages, through the EventKind::parse method, however most System Common and System Realtime messages are unsupported.

Modules

number
Special-length integers used by the MIDI standard.

Structs

Error
Represents an error parsing an SMF file or MIDI stream.

Event
Represents a fully parsed track event, with delta time.

Header
A MIDI file header.

Smf
Represents a Standard Midi File (.mid and .midi files). Yields TrackRepr on a Vec, allowing for customization on what is stored in a track. References an outside byte array.

SmpteTime
Encodes an SMPTE time of the day.

TrackIter
Allows deferring track parsing for later, on a per-event basis.

Enums

ErrorKind
The type of error that occurred while parsing.

EventKind
Represents the different kinds of events.

Format
The different formats an SMF file can be.

Fps

MetaMessage

MidiMessage
Represents a MIDI message, not an event.

Timing
The timing for an SMF file. This can be in ticks/beat or ticks/second.

Traits

TrackRepr
Allows for customization on how tracks are stored in memory.